workshop/app/back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "A modern Python backend for the StarWarsKubernetes project."
authors = ["GitHub Copilot"]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlmodel",
    "psycopg2-binary",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "bcrypt==3.2.0",
    "python-dotenv",
    "httpx",
    "python-multipart",
    "pytest",
    "google-generativeai",
    "Pillow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
